moved upload_profile_photo from app.py
