+++ app.py
@@ -25,7 +25,8 @@ from routes import (
     change_email_pb,
     change_username_pb,
     upload_pb,
-    edit_image_pb
+    edit_image_pb,
+    upload_profile_photo_pb
 )
 from auth.auth import login_is_required, role_required
 
@@ -65,6 +66,7 @@ app.register_blueprint(change_email_pb)
 app.register_blueprint(change_username_pb)
 app.register_blueprint(upload_pb)
 app.register_blueprint(edit_image_pb)
+app.register_blueprint(upload_profile_photo_pb)
 
 flow = Flow.from_client_secrets_file(
     client_secrets_file=client_secrets_file,
@@ -182,56 +184,6 @@ def upload_images():
             flash('Invalid file type or no file selected.', 'danger')
 
     return redirect(url_for('profile'))
-
-@app.route('/upload_profile_photo', methods=['POST'])
-def upload_profile_photo():
-    if 'username' not in session:
-        flash('Please log in to update your profile photo.', 'danger')
-        return redirect(url_for('login'))
-    
-    username = session['username']
-    user = get_user_by_username(username)
-    if not user:
-        flash('User not found.', 'danger')
-        return redirect(url_for('login'))
-    
-    if 'profile_photo' not in request.files:
-        flash('No file selected.', 'danger')
-        return redirect(url_for('profile'))
-    
-    file = request.files['profile_photo']
-    
-    if file.filename == '':
-        flash('No file selected.', 'danger')
-        return redirect(url_for('profile'))
-    
-    if file and allowed_file(file.filename):
-        filename = f"{username}_profile.{file.filename.rsplit('.', 1)[1].lower()}"
-        upload_folder = os.path.join(app.config['UPLOAD_FOLDER'], 'profile')
-
-        if not os.path.exists(upload_folder):
-            os.makedirs(upload_folder)  # Create missing directories
-        
-        filepath = os.path.join(upload_folder, filename)
-        
-        # Remove old profile photo if it exists
-        if user.get('profile_photo'):
-            old_filepath = os.path.join(app.config['UPLOAD_FOLDER'], 'profile', user['profile_photo'])
-            if os.path.exists(old_filepath):
-                os.remove(old_filepath)
-        
-        # Save the new file
-        file.save(filepath)
-        
-        # Update user record in database with the new profile photo filename
-        userdatahandler.update_profile_photo(username, filename)
-        
-        flash('Profile photo updated successfully!', 'success')
-    else:
-        flash('Invalid file type. Only jpg, jpeg, png, and gif files are allowed.', 'danger')
-    
-    return redirect(url_for('profile'))
-
  
 @app.route('/audio/<filename>')
 def serve_audio(filename):
